[Feat] 🍺 영업점 목록 api 정렬 구현

+++ src/main/java/com/hanaro/schedule_hanaro/customer/dto/response/BranchDetailResponse.java
@@ -36,7 +36,7 @@ public record BranchDetailResponse(
 	List<Integer> waitTime,
 
 	@JsonProperty("distance")
-	long distance
+	double distance
 ) {
 	public static BranchDetailResponse of(Long id, String branchName, String xPosition, String yPosition,
 		String address) {
